% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fasta2df}
\alias{fasta2df}
\title{Convert a FASTA file to a data.frame}
\usage{
fasta2df(fasta_path)
}
\arguments{
\item{fasta_path}{a path to the fasta file.}
}
\value{
It returns a data.frame that contains information of guide RNA library.
}
\description{
Convert a FASTA file to a data.frame
}
\examples{
library(CB2)

fa_path <- system.file(
 "extdata",
 "toydata",
 "small_sample.fasta",
 package = "CB2"
) 
cat(readLines(fa_path)[1:5], sep="\n") # To see the head of the file.

fasta2df(fa_path)
}
