% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_ab}
\alias{fit_ab}
\title{A C++ function to perform a parameter estimation for the sgRNA-level test. 
It will estimate two different parameters `phat` and `vhat`,
and we believe input count data follows the beta-binomial distribution.}
\usage{
fit_ab(xvec, nvec)
}
\arguments{
\item{xvec}{a matrix contains sgRNA read counts.}

\item{nvec}{a vector contains the library size.}
}
\description{
A C++ function to perform a parameter estimation for the sgRNA-level test. 
It will estimate two different parameters `phat` and `vhat`,
and we believe input count data follows the beta-binomial distribution.
}
