% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_CPM}
\alias{get_CPM}
\title{A function to normalize sgRNA read counts.}
\usage{
get_CPM(sgcount)
}
\arguments{
\item{sgcount}{The input data.frame contains read countss of sgRNAs for each sample

A function to calculate the CPM (Counts Per Million) (required)}
}
\value{
a normalized CPM data.frame will be returned
}
\description{
A function to normalize sgRNA read counts.
}
\examples{
library(CB2)
data(Evers_CRISPRn_RT112)
get_CPM(Evers_CRISPRn_RT112$count)

}
