% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_corr_heatmap}
\alias{plot_corr_heatmap}
\title{A function to show a heatmap sgRNA-level corrleations of the NGS samples.}
\usage{
plot_corr_heatmap(sgcount, df_design, cor_method = "pearson")
}
\arguments{
\item{sgcount}{The input matrix contains read counts of sgRNAs for each sample.}

\item{df_design}{The table contains a study design.}

\item{cor_method}{A string parameter of the correlation measure. One of the three - "pearson", "kendall", or "spearman" will be the string.}
}
\value{
A pheatmap object contains the correlation heatmap

library(CB2)
data(Evers_CRISPRn_RT112)
plot_corr_heatmap(Evers_CRISPRn_RT112$count, Evers_CRISPRn_RT112$design)
}
\description{
A function to show a heatmap sgRNA-level corrleations of the NGS samples.
}
