% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_count_distribution}
\alias{plot_count_distribution}
\title{A function to plot read count distribution.}
\usage{
plot_count_distribution(sgcount, df_design)
}
\arguments{
\item{sgcount}{The input matrix contains read counts of sgRNAs for each sample.}

\item{df_design}{The data.frame contains a study design.}
}
\value{
A ggplot object contains a read count distribution plot for `sgcount`.
}
\description{
A function to plot read count distribution.
}
\examples{
library(CB2)
data(Evers_CRISPRn_RT112)
cpm <- get_CPM(Evers_CRISPRn_RT112$count)
plot_count_distribution(cpm, Evers_CRISPRn_RT112$design)

}
