% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep_row}
\alias{rep_row}
\title{Repeat a row}
\usage{
rep_row(x, n)
}
\arguments{
\item{x}{A vector to repeat}

\item{n}{Number of repetitions}
}
\value{
A matrix with `n` copies of `x`
}
\description{
Repeat a row
}
