% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sanson_CRISPRn_A375}
\alias{Sanson_CRISPRn_A375}
\title{A benchmark CRISPRn pooled screen data from Sanson et al.}
\format{
The data object is a list and contains below information:
\describe{
  \item{count}{The count matrix from Sanson et al.'s paper and contains the CRISPRn screening result using A375 cell-line. 
    It contains a sample of plasimd, and three biological replicates after three weeks.}
  \item{egenes}{The list of 1,580 essential genes used in Sanson et al.'s study.}
  \item{ngenes}{The list of 927 non-essential genes used in Sanson et al.'s study.}
  \item{design}{The data.frame contains study design.}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/30575746/}
}
\usage{
data(Sanson_CRISPRn_A375)
}
\description{
A benchmark CRISPRn pooled screen data from Sanson et al.
}
\keyword{datasets}
