% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_count_and_design}
\alias{join_count_and_design}
\title{A function to join a count table and a design table.}
\usage{
join_count_and_design(sgcount, df_design)
}
\arguments{
\item{sgcount}{The input matrix contains read counts of sgRNAs for each sample.}

\item{df_design}{The table contains a study design.}
}
\value{
A tidy and joined table will be returned.
}
\description{
A function to join a count table and a design table.
}
