% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_dotplot}
\alias{plot_dotplot}
\title{A function to visualze dotplots for a gene.}
\usage{
plot_dotplot(sgcount, df_design, gene)
}
\arguments{
\item{sgcount}{The input matrix contains read counts of sgRNAs for each sample.}

\item{df_design}{The data.frame contains a study design.}

\item{gene}{The gene to be visualzed.}
}
\value{
A ggplot object contains dotplots of sgRNA read counts for a gene.
}
\description{
A function to visualze dotplots for a gene.
}
\examples{
library(CB2)
data(Evers_CRISPRn_RT112)
plot_dotplot(get_CPM(Evers_CRISPRn_RT112$count), Evers_CRISPRn_RT112$design, "RPS7")

}
