% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table2fa}
\alias{table2fa}
\title{Convert a CSV/TSV format file to a FASTA file.}
\usage{
table2fa(input_path, output_path, seq_col = 1, gene_col = 2)
}
\arguments{
\item{input_path}{A path to a CSV/TSV (comma/tab-separated values) format file.}

\item{output_path}{A path to the output file.}

\item{seq_col}{A column index or column name whose column contains guide RNA sequences.}

\item{gene_col}{A column index or column name whose column contains gene names. 
The type of both `seq_col` and `gene_col` have to be the same.}
}
\value{

}
\description{
Convert a CSV/TSV format file to a FASTA file.
}
